requirements.txt
boto3==1.37.5
botocore==1.37.5
Django==5.1.6
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
PyMySQL==1.1.1
python-dotenv==1.0.1
